+++ datasift/__init__.py
@@ -44,8 +44,8 @@ more details.
 
 __author__ = "opensource@datasift.com"
 __status__ = "stable"
-__version__ = "2.6.0"
-__date__ = "25th Nov 2015"
+__version__ = "2.10.0"
+__date__ = "2nd Nov 2016"
 
 # -----------------------------------------------------------------------------
 # Module constants